refactor: moved command types to typings
