+++ src/modules/utils.ts
@@ -8,6 +8,7 @@ import {
     Colors,
     CommandInteraction,
     ComponentType,
+    ContextMenuCommandBuilder,
     EmbedBuilder,
     Guild,
     GuildEmoji,
@@ -16,10 +17,21 @@ import {
     RepliableInteraction,
     Routes,
     Serialized,
+    SlashCommandBuilder,
+    SlashCommandSubcommandBuilder,
+    SlashCommandSubcommandGroupBuilder,
     StringSelectMenuBuilder,
     TextBasedChannel,
     User,
 } from 'discord.js';
+import type {
+    ContextCommand,
+    InteractionCommand,
+    MessageCommand,
+    SlashCommand,
+    SlashSubcommand,
+    SlashSubcommandGroup,
+} from '@typings/commands';
 
 function strip(text: string, char: string) {
     return text.replaceAll(new RegExp(`^${char}+|${char}+$`, 'g'), '');
@@ -383,4 +395,42 @@ export async function* fetch_history(
         }
         amount -= i;
     }
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isSlashSubcommand(obj: any): obj is SlashSubcommand {
+    return obj && obj.data instanceof SlashCommandSubcommandBuilder &&
+        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
+        obj.execute.length <= 2;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isSlashSubcommandGroup(obj: any): obj is SlashSubcommandGroup {
+    return obj && obj.data instanceof SlashCommandSubcommandGroupBuilder &&
+        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
+        obj.subcommands && obj.execute.length <= 2;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isSlashCommand(obj: any): obj is SlashCommand {
+    return obj && obj.data instanceof SlashCommandBuilder &&
+        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
+        obj.execute.length <= 2;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isContextCommand(obj: any): obj is ContextCommand {
+    return obj && obj.data instanceof ContextMenuCommandBuilder &&
+        typeof obj.execute === 'function' && obj.execute.length <= 2;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isInteractionCommand(obj: any): obj is InteractionCommand {
+    return isSlashCommand(obj) || isContextCommand(obj);
+}
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export function isMessageCommand(obj: any): obj is MessageCommand {
+    return obj && typeof obj.name === 'string' && typeof obj.admin === 'boolean' &&
+        typeof obj.desc === 'string' && typeof obj.execute === 'function' && obj.execute.length <= 3;
 }
