+++ src/typings/README.md
@@ -1,0 +1,9 @@
+# Source Folder
+
+This folder contains all custom typings that are needed for TypeScript.
+
+## Files
+
+- [client.d.ts](client.d.ts): Custom typings for the discord.js client. This is used to extend the client with custom
+  properties and methods.
+- [commands.d.ts](commands.d.ts): Typings for slash commands, message commands, and helpers.
