+++ src/typings/commands.d.ts
@@ -1,4 +1,4 @@
-import {
+import type {
     AnySelectMenuInteraction,
     ButtonInteraction,
     ChatInputCommandInteraction,
@@ -7,50 +7,12 @@ import {
     Message,
     ModalSubmitInteraction,
     SharedNameAndDescription,
-    SlashCommandBuilder,
     SlashCommandOptionsOnlyBuilder,
     SlashCommandSubcommandBuilder,
     SlashCommandSubcommandGroupBuilder,
     SlashCommandSubcommandsOnlyBuilder,
 } from 'discord.js';
 import type { Cache } from '@modules/database';
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isSlashSubcommand(obj: any): obj is SlashSubcommand {
-    return obj && obj.data instanceof SlashCommandSubcommandBuilder &&
-        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
-        obj.execute.length <= 2;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isSlashSubcommandGroup(obj: any): obj is SlashSubcommandGroup {
-    return obj && obj.data instanceof SlashCommandSubcommandGroupBuilder &&
-        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
-        obj.subcommands && obj.execute.length <= 2;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isSlashCommand(obj: any): obj is SlashCommand {
-    return obj && obj.data instanceof SlashCommandBuilder &&
-        typeof obj.desc === 'string' && typeof obj.execute === 'function' &&
-        obj.execute.length <= 2;
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isContextCommand(obj: any): obj is ContextCommand {
-    return obj && obj.data instanceof ContextMenuCommandBuilder &&
-        typeof obj.execute === 'function' && obj.execute.length <= 2;
-}
-
-export function isInteractionCommand(obj: unknown): obj is InteractionCommand {
-    return isSlashCommand(obj) || isContextCommand(obj);
-}
-
-// eslint-disable-next-line @typescript-eslint/no-explicit-any
-export function isMessageCommand(obj: any): obj is MessageCommand {
-    return obj && typeof obj.name === 'string' && typeof obj.admin === 'boolean' &&
-        typeof obj.desc === 'string' && typeof obj.execute === 'function' && obj.execute.length <= 3;
-}
 
 export type CommandFile = {
     name: string;
